% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clif_c_ad.R
\name{clif_ad}
\alias{clif_ad}
\title{Title CLIF-C-AD score calculation}
\usage{
clif_ad(age, creatinine, inr, wbc, sodium, digits = 0, unit = "US")
}
\arguments{
\item{age}{Numeric vector of age in years.}

\item{creatinine}{Numeric vector of creatinine in either µmol/L or mg/dL.}

\item{inr}{Numeric vector of International normalized ratio (INR)}

\item{wbc}{Numeric vector of white blood cell count (x 10⁹ cells/L)}

\item{sodium}{Numeric vector of sodium (mmol/L)}

\item{digits}{Option for specifying digits for rounding (default = 0)}

\item{unit}{Option specifying if creatinine is using SI or US units. If the
option i set to SI, creatinine is automatically converted to mg/dL before
calculating the CLIF-C-AD score.}
}
\value{
Numeric vector of CLIF-C-AD score
}
\description{
Title CLIF-C-AD score calculation
}
\details{
Additional details...
}
\examples{
clif_ad(age = 50, creatinine = 2, inr = 2, wbc = 20, sodium = 137, unit = "US")
# 69

# Or simply:
clif_ad(50, 2, 2, 20, 137)
# 69
 
# If your creatinine follows SI units, then specify it before calculating:
clif_ad(age = 50, creatinine = 90, inr = 2, wbc = 20, sodium = 137, unit = "SI")
# 65

# The function is fully vectorized and should handle multiple values:
clif_ad(age = c(50, 70, 22), creatinine = c(90, 120, 70),
 inr = c(2, 1.8, 2.2), wbc = c(20, 60, 11),
  sodium = c(137, 122, 140), unit = "SI")
 # 65, 88, 55
}
\references{
Jalan, R., Pavesi, M., Saliba, F., Amorós, A., Fernandez, J.,
Holland-Fischer, P., Sawhney, R., Mookerjee, R., Caraceni, P., Moreau, R.,
Ginès, P., Durand, F., Angeli, P., Alessandria, C., Laleman, W., Trebicka,
J., Samuel, D., Zeuzem, S., Gustot, T., Gerbes, A. L., Wendon, J., Bernardi,
M., Arroyo, V., CANONIC study investigators of the EASL-CLIF Consortium.
The CLIF Consortium Acute Decompensation score (CLIF-C ADs) for prognosis of hospitalised cirrhotic patients without acute-on-chronic liver failure.
J. Hepatol. 2015, 62 (4), 831–840. DOI: 10.1016/j.jhep.2014.11.012
}
