% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liverscorer-package.R
\docType{package}
\name{liverscoreR-package}
\alias{liverscoreR}
\alias{liverscoreR-package}
\title{liverscoreR: Collection of Clinical Scores used in Hepatology}
\description{
Collection of clinical scores used in hepatology.
}
\author{
\strong{Maintainer}: Søren Damsholt \email{soerdm@clin.au.dk}

Authors:
\itemize{
  \item Søren Damsholt \email{soerdm@clin.au.dk}
}

}
\keyword{internal}
