% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meldna_score.R
\name{meldna}
\alias{meldna}
\title{Title Model for End-stage Liver Disease - Sodium score}
\usage{
meldna(
  creatinine,
  bilirubin,
  inr,
  sodium,
  unit = "US",
  dialysis = "no",
  creatinine_conversion_factor = 0.0113,
  bilirubin_conversion_factor = 0.0584,
  creatinine_threshold_si = 200,
  bilirubin_threshold_si = 360,
  creatinine_threshold_us = 2.26,
  bilirubin_threshold_us = 21.05,
  sodium_upper_threshold = 160,
  sodium_lower_threshold = 100,
  inr_threshold = 3
)
}
\arguments{
\item{creatinine}{Serum creatinine in either mg/dL (US units) or μmol/L (SI units).}

\item{bilirubin}{Serum total bilirubin in either mg/dL (US units) or μmol/L (SI units).}

\item{inr}{International normalized ratio of coagulation factor II, VII and X.}

\item{sodium}{Serum sodium in mmol/L.}

\item{unit}{Parameter setting if the measurements are based on "US" units (mg/dL) or SI units (μmol/L). Default value is "US". If the setting is changed to "SI", creatinine and bilirubin will be converted to mg/dL by multiplying their respective conversion factor.}

\item{dialysis}{Parameter setting if the patient is on hemodialysis at the time of measurement (default = "no"), if yes, then creatinine is hard-coded to 4.0 mg/dL before calculation.}

\item{creatinine_conversion_factor}{Factor multiplied to creatinine to convert from μmol/L to mg/dL (default = 0.0113).}

\item{bilirubin_conversion_factor}{Factor multiplied to bilirubin to convert from μmol/L to mg/dL (default = 0.0584).}

\item{creatinine_threshold_si}{Upper threshold for high creatinine (μmol/L), causing a warning if input is above (default = 200).}

\item{bilirubin_threshold_si}{Upper threshold for high bilirubin (μmol/L), causing a warning if input is above (default = 360).}

\item{creatinine_threshold_us}{Upper threshold for high creatinine (mg/dL), causing a warning if input is higher (default = 2.26).}

\item{bilirubin_threshold_us}{Upper threshold for high bilirubin (mg/dL), causing a warning if input is higher (default = 21.05).}

\item{sodium_upper_threshold}{Upper threshold for extremely high sodium measurements (default = 100)}

\item{sodium_lower_threshold}{Lower threshold for extremely low sodium levels (default = 160)}

\item{inr_threshold}{Upper threshold for INR, causing a warning if input is higher (default = 3.0).}
}
\value{
A numeric vector of MELD-Na scores rounded to the nearest integer
}
\description{
Calculates the Model for End-Stage Liver Disease incorporating serum sodium (MELD-Na) score.
The calculation follows the 2016 MELD-Na formulation
}
\details{
The MELD-Na score is calculated from S-creatinine, S-bilirubin, INR and S-sodium.
Creatinine, bilirubin and INR are bounded so values < 1 are set to 1. Sodium is bounded between
125 and 137 Both US and SI units are applicable for this function, where SI units are automatically converted to mg/dl before calculation.
The MELD-Na calculation follows two steps:
\enumerate{
\item MELD(i) is calculated as:
\deqn{
     MELD(i) = 10 \times round(
       0.957 \times \log(Creatinine) + 
       0.378 \times \log(Bilirubin) + 
       1.120 \times \log(INR) + 0.643, digits = 10)}
\item If MELD(i) > 11, the sodium adjustment is applied:
\deqn{
     MELDNa = MELD(i) + 1.32(137-Na)-
       0.033\,MELD(i)(137-Na)
     }
}

Extreme outliers will flag a warning from the package, these can manually be set higher or lower by the user.
If one or more required input values are missing for an observation, the corresponding MELD-Na score is returned as "Na"

The original MELD-Na model was described by Kim et al. (2008).
This function implements the subsequent 2016 MELD-Na formulation
adopted for liver allocation in the United States.
}
\examples{
  meldna(creatinine = 1, bilirubin = 1, inr = 1, sodium = 125, unit = "US", dialysis = "no") 
  # 6. 
  
  # This can be simplified to:
  meldna(1, 1, 1, 125)
  
  # SI units can be applied and are automatically converted to mg/dL
  meldna(120, 80, 2, 125, unit = "SI", dialysis = "no")
  # 30
  
  # If the patient has received haemodialysis, creatinine is hardcoded to 4.0. 
  # This setting can be enabled by setting "dialysis" to "yes"
  meldna(120, 80, 2, 125, unit = "SI", dialysis = "yes")
  # 36
  
}
\references{
Kim WR, Biggins SW, Kremers WK, Wiesner RH, Kamath PS, Benson JT,
Edwards E, Therneau TM. Hyponatremia and mortality among patients
on the liver-transplant waiting list. N Engl J Med.
2008;359(10):1018-1026. doi:10.1056/NEJMoa0801209.

United Network for Organ Sharing (UNOS). Policy and system changes
effective January 11, 2016, adding serum sodium to MELD calculation.
}
\seealso{
\code{\link[=meld]{meld()}} for the original MELD score
}
